extract pipeline job handling to service
