+++ api-server/api/githubwebhook/dummy.go
@@ -1,0 +1,1 @@
+package githubwebhook
